# src/main/java/com/gestao/dto/LoginDTO.java
package com.gestao.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@Schema(description = "DTO para login de usuário")
public class LoginDTO {

    @NotBlank(message = "Username é obrigatório")
    @Schema(description = "Nome de usuário", example = "joao_silva")
    private String username;

    @NotBlank(message = "Senha é obrigatória")
    @Schema(description = "Senha do usuário", example = "senha123")
    private String senha;
}